feat: change compozition favorite 30.07.23 11:13
