+++ chebProSport/src/pages/favorite.tsx
@@ -1,20 +1,10 @@
 import React from 'react';
-import {
-  Container,
-  Grid,
-  Link,
-  Typography,
-  Card,
-  Box,
-  CardContent,
-  CardActions,
-} from '@mui/material';
+import { Container, Grid, Link, Typography } from '@mui/material';
 import Breadcrumbs from '@mui/material/Breadcrumbs';
-import BasketBtn from '../components/catalog/item/basketBtn';
 import NavigateNextIcon from '@mui/icons-material/NavigateNext';
 import HighlightOffIcon from '@mui/icons-material/HighlightOff';
-import CardCatalog from '../data/catalog';
-import { useGlobalContext } from '../store/AppStore';
+
+import CardFavorite from './cardFavorite';
 
 const breadcrumbs = [
   <Link underline='hover' key='1' color='inherit' href='/'>
@@ -25,8 +15,6 @@ const breadcrumbs = [
   </Typography>,
 ];
 const Favorite = () => {
-  const { count } = useGlobalContext();
-
   return (
     <Container maxWidth={'md'}>
       <Grid container spacing={2}>
@@ -58,66 +46,7 @@ const Favorite = () => {
             flexWrap: 'wrap',
           }}
         >
-          {count.map((item) => {
-            const cardObj = CardCatalog.find((card) => card.id === item);
-
-            return (
-              <>
-                <Card
-                  sx={{
-                    maxWidth: '245px',
-                    height: 'auto',
-                    flexGrow: '1',
-                    display: 'flex',
-                    flexDirection: 'column',
-                  }}
-                  key={item}
-                >
-                  <Box
-                    sx={{
-                      display: 'flex',
-                      justifyContent: 'center',
-                      height: '190px',
-                      cursor: 'pointer',
-                    }}
-                  >
-                    <img
-                      src={cardObj?.image}
-                      alt='Catalog Card'
-                      style={{ paddingTop: '15px' }}
-                    />
-                  </Box>
-                  <CardContent sx={{ padding: '5px 5px 0px 10px' }}>
-                    <Typography
-                      gutterBottom
-                      variant='body1'
-                      component='span'
-                      sx={{ mr: '10px' }}
-                    >
-                      {cardObj?.title}
-                    </Typography>
-                    <Typography
-                      variant='body1'
-                      component={'span'}
-                      color='secondary'
-                    >
-                      {cardObj?.price}
-                    </Typography>
-                    <Typography
-                      variant='body2'
-                      component={'div'}
-                      sx={{ color: '#47B315', mt: '10px' }}
-                    >
-                      в наличии {cardObj?.score}
-                    </Typography>
-                  </CardContent>
-                  <CardActions>
-                    <BasketBtn />
-                  </CardActions>
-                </Card>
-              </>
-            );
-          })}
+          <CardFavorite />
         </Grid>
       </Grid>
     </Container>
